requirements.txt
Django==4.0
django-debug-toolbar==3.2.4
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
Pillow==10.3.0
django-environ==0.11.2
djangorestframework==3.15.1